+++ okhttp/src/jvmMain/kotlin/okhttp3/internal/connection/ConnectInterceptor.kt
@@ -29,7 +29,7 @@ object ConnectInterceptor : Interceptor {
   @Throws(IOException::class)
   override fun intercept(chain: Interceptor.Chain): Response {
     val realChain = chain as RealInterceptorChain
-    val exchange = realChain.call.initExchange()
+    val exchange = realChain.call.initExchange(realChain)
     val connectedChain = realChain.copy(exchange = exchange)
     return connectedChain.proceed(realChain.request)
   }
