+++ okhttp/src/jvmMain/kotlin/okhttp3/internal/connection/ExchangeFinder.kt
@@ -16,353 +16,33 @@
 package okhttp3.internal.connection
 
 import java.io.IOException
-import java.net.Socket
-import okhttp3.Address
-import okhttp3.HttpUrl
-import okhttp3.OkHttpClient
-import okhttp3.Route
-import okhttp3.internal.canReuseConnectionFor
-import okhttp3.internal.closeQuietly
-import okhttp3.internal.http.ExchangeCodec
-import okhttp3.internal.http.RealInterceptorChain
-import okhttp3.internal.http2.ConnectionShutdownException
-import okhttp3.internal.http2.ErrorCode
-import okhttp3.internal.http2.StreamResetException
 
-/**
- * Attempts to find the connections for an exchange and any retries that follow. This uses the
- * following strategies:
- *
- *  1. If the current call already has a connection that can satisfy the request it is used. Using
- *     the same connection for an initial exchange and its follow-ups may improve locality.
- *
- *  2. If there is a connection in the pool that can satisfy the request it is used. Note that it is
- *     possible for shared exchanges to make requests to different host names! See
- *     [RealConnection.isEligible] for details.
- *
- *  3. If there's no existing connection, make a list of routes (which may require blocking DNS
- *     lookups) and attempt a new connection them. When failures occur, retries iterate the list of
- *     available routes.
- *
- * If the pool gains an eligible connection while DNS, TCP, or TLS work is in flight, this finder
- * will prefer pooled connections. Only pooled HTTP/2 connections are used for such de-duplication.
- *
- * It is possible to cancel the finding process.
- *
- * Instances of this class are not thread-safe. Each instance is thread-confined to the thread
- * executing [call].
- */
-class ExchangeFinder(
-  private val client: OkHttpClient,
-  internal val address: Address,
-  private val call: RealCall,
-  private val chain: RealInterceptorChain,
+internal class ExchangeFinder(
+  private val routePlanner: RoutePlanner
 ) {
-  private val connectionPool = client.connectionPool.delegate
-  private val eventListener = call.eventListener
-  private val doExtensiveHealthChecks = chain.request.method != "GET"
-
-  private var routeSelection: RouteSelector.Selection? = null
-  private var routeSelector: RouteSelector? = null
-  private var refusedStreamCount = 0
-  private var connectionShutdownCount = 0
-  private var otherFailureCount = 0
-  private var nextRouteToTry: Route? = null
-
-  fun find(): ExchangeCodec {
+  fun find(): RealConnection {
     var firstException: IOException? = null
     while (true) {
-      if (call.isCanceled()) throw IOException("Canceled")
+      if (routePlanner.isCanceled()) throw IOException("Canceled")
 
       try {
-        val plan = plan()
-        when (plan) {
-          is ReusePlan -> return plan.connection.newCodec(client, chain)
-
-          is ConnectPlan -> {
-            plan.connect()
-
-            val failure = plan.failure
-            if (failure != null) throw failure
-
-            val result = plan.handleSuccess()
-            return result.newCodec(client, chain)
-          }
+        val plan = routePlanner.plan()
+        if (!plan.isConnected) {
+          plan.connect()
         }
+        return plan.handleSuccess()
       } catch (e: IOException) {
-        trackFailure(e)
+        routePlanner.trackFailure(e)
 
         if (firstException == null) {
           firstException = e
         } else {
           firstException.addSuppressed(e)
         }
-        if (!retryAfterFailure()) {
+        if (!routePlanner.retryAfterFailure()) {
           throw firstException
         }
       }
     }
-  }
-
-  /**
-   * Returns a healthy connection to host a new stream. This prefers the existing connection if it
-   * exists, then the pool, finally building a new connection.
-   *
-   * This checks for cancellation before each blocking operation.
-   */
-  @Throws(IOException::class)
-  private fun plan(): Plan {
-    val reuseCallConnection = planReuseCallConnection()
-    if (reuseCallConnection != null) return reuseCallConnection
-
-    // We need a new connection. Give it fresh stats.
-    refusedStreamCount = 0
-    connectionShutdownCount = 0
-    otherFailureCount = 0
-
-    // Attempt to get a connection from the pool.
-    val pooled1 = planReusePooledConnection()
-    if (pooled1 != null) return pooled1
-
-    // Do blocking calls to plan a route for a new connection.
-    val connect = planConnect()
-
-    // Now that we have a set of IP addresses, make another attempt at getting a connection from
-    // the pool. We have a better chance of matching thanks to connection coalescing.
-    val pooled2 = planReusePooledConnection(connect.connection, connect.routes)
-    if (pooled2 != null) return pooled2
-
-    return connect
-  }
-
-  /**
-   * Returns the connection already attached to the call if it's eligible for a new exchange.
-   *
-   * If the call's connection exists and is eligible for another exchange, it is returned. If it
-   * exists but cannot be used for another exchange, it is closed and this returns null.
-   */
-  private fun planReuseCallConnection(): ReusePlan? {
-    // This may be mutated by releaseConnectionNoEvents()!
-    val candidate = call.connection ?: return null
-
-    // Make sure this connection is healthy & eligible for new exchanges. If it's no longer needed
-    // then we're on the hook to close it.
-    val healthy = candidate.isHealthy(doExtensiveHealthChecks)
-    val toClose: Socket? = synchronized(candidate) {
-      when {
-        !healthy -> {
-          candidate.noNewExchanges = true
-          call.releaseConnectionNoEvents()
-        }
-        candidate.noNewExchanges || !sameHostAndPort(candidate.route().address.url) -> {
-          call.releaseConnectionNoEvents()
-        }
-        else -> null
-      }
-    }
-
-    // If the call's connection wasn't released, reuse it. We don't call connectionAcquired() here
-    // because we already acquired it.
-    if (call.connection != null) {
-      check(toClose == null)
-      return ReusePlan(candidate)
-    }
-
-    // The call's connection was released.
-    toClose?.closeQuietly()
-    eventListener.connectionReleased(call, candidate)
-    return null
-  }
-
-  /** Plans to make a new connection by deciding which route to try next. */
-  @Throws(IOException::class)
-  private fun planConnect(): ConnectPlan {
-    // Use a route from a preceding coalesced connection.
-    val localNextRouteToTry = nextRouteToTry
-    if (localNextRouteToTry != null) {
-      nextRouteToTry = null
-      return ConnectPlan(localNextRouteToTry)
-    }
-
-    // Use a route from an existing route selection.
-    val existingRouteSelection = routeSelection
-    if (existingRouteSelection != null && existingRouteSelection.hasNext()) {
-      return ConnectPlan(existingRouteSelection.next())
-    }
-
-    // Decide which proxy to use, if any. This may block in ProxySelector.select().
-    var newRouteSelector = routeSelector
-    if (newRouteSelector == null) {
-      newRouteSelector = RouteSelector(address, call.client.routeDatabase, call, eventListener)
-      routeSelector = newRouteSelector
-    }
-
-    // List available IP addresses for the current proxy. This may block in Dns.lookup().
-    if (!newRouteSelector.hasNext()) throw IOException("exhausted all routes")
-    val newRouteSelection = newRouteSelector.next()
-    routeSelection = newRouteSelection
-
-    if (call.isCanceled()) throw IOException("Canceled")
-
-    return ConnectPlan(newRouteSelection.next(), newRouteSelection.routes)
-  }
-
-  /**
-   * Returns a plan to reuse a pooled connection, or null if the pool doesn't have a connection for
-   * this address.
-   *
-   * If [connectionToReplace] is non-null, this will swap it for a pooled connection if that
-   * pooled connection uses HTTP/2. That results in fewer sockets overall and thus fewer TCP slow
-   * starts.
-   */
-  private fun planReusePooledConnection(
-    connectionToReplace: RealConnection? = null,
-    routes: List<Route>? = null,
-  ): ReusePlan? {
-    val result = connectionPool.callAcquirePooledConnection(
-      doExtensiveHealthChecks = doExtensiveHealthChecks,
-      address = address,
-      call = call,
-      routes = routes,
-      requireMultiplexed = connectionToReplace != null && !connectionToReplace.isNew
-    ) ?: return null
-
-    // If we coalesced our connection, remember the replaced connection's route. That way if the
-    // coalesced connection later fails we don't waste a valid route.
-    if (connectionToReplace != null) {
-      nextRouteToTry = connectionToReplace.route()
-      if (!connectionToReplace.isNew) {
-        connectionToReplace.socket().closeQuietly()
-      }
-    }
-
-    eventListener.connectionAcquired(call, result)
-    return ReusePlan(result)
-  }
-
-  internal sealed class Plan
-
-  /** Reuse an existing connection. */
-  internal class ReusePlan(
-    val connection: RealConnection,
-  ) : Plan()
-
-  /** Establish a new connection. */
-  internal inner class ConnectPlan(
-    route: Route,
-    val routes: List<Route>? = null,
-  ) : Plan() {
-    val connection = RealConnection(client.taskRunner, connectionPool, route)
-    var complete = false
-    var failure: IOException? = null
-
-    fun connect() {
-      // Tell the call about the connecting call so async cancels work.
-      call.connectionsToCancel += connection
-      try {
-        connection.connect(
-          connectTimeout = chain.connectTimeoutMillis,
-          readTimeout = chain.readTimeoutMillis,
-          writeTimeout = chain.writeTimeoutMillis,
-          pingIntervalMillis = client.pingIntervalMillis,
-          connectionRetryEnabled = client.retryOnConnectionFailure,
-          call = call,
-          eventListener = eventListener,
-        )
-      } catch (e: IOException) {
-        failure = e
-      } finally {
-        call.connectionsToCancel -= connection
-        complete = true
-      }
-    }
-
-    /** Returns the connection to use, which might be different from [connection]. */
-    fun handleSuccess(): RealConnection {
-      call.client.routeDatabase.connected(connection.route())
-
-      // If we raced another call connecting to this host, coalesce the connections. This makes for
-      // 3 different lookups in the connection pool!
-      val pooled3 = planReusePooledConnection(connection, routes)
-      if (pooled3 != null) return pooled3.connection
-
-      synchronized(connection) {
-        connectionPool.put(connection)
-        call.acquireConnectionNoEvents(connection)
-      }
-
-      eventListener.connectionAcquired(call, connection)
-      return connection
-    }
-  }
-
-  fun trackFailure(e: IOException) {
-    if (e is StreamResetException && e.errorCode == ErrorCode.REFUSED_STREAM) {
-      refusedStreamCount++
-    } else if (e is ConnectionShutdownException) {
-      connectionShutdownCount++
-    } else {
-      otherFailureCount++
-    }
-  }
-
-  /**
-   * Returns true if the current route has a failure that retrying could fix, and that there's
-   * a route to retry on.
-   */
-  fun retryAfterFailure(): Boolean {
-    if (refusedStreamCount == 0 && connectionShutdownCount == 0 && otherFailureCount == 0) {
-      return false // Nothing to recover from.
-    }
-
-    if (nextRouteToTry != null) {
-      return true
-    }
-
-    val retryRoute = retryRoute()
-    if (retryRoute != null) {
-      // Lock in the route because retryRoute() is racy and we don't want to call it twice.
-      nextRouteToTry = retryRoute
-      return true
-    }
-
-    // If we have a routes left, use 'em.
-    if (routeSelection?.hasNext() == true) return true
-
-    // If we haven't initialized the route selector yet, assume it'll have at least one route.
-    val localRouteSelector = routeSelector ?: return true
-
-    // If we do have a route selector, use its routes.
-    return localRouteSelector.hasNext()
-  }
-
-  /**
-   * Return the route from the current connection if it should be retried, even if the connection
-   * itself is unhealthy. The biggest gotcha here is that we shouldn't reuse routes from coalesced
-   * connections.
-   */
-  private fun retryRoute(): Route? {
-    if (refusedStreamCount > 1 || connectionShutdownCount > 1 || otherFailureCount > 0) {
-      return null // This route has too many problems to retry.
-    }
-
-    val connection = call.connection ?: return null
-
-    synchronized(connection) {
-      if (connection.routeFailureCount != 0) return null
-      if (!connection.noNewExchanges) return null // This route is still in use.
-      if (!connection.route().address.url.canReuseConnectionFor(address.url)) return null
-      return connection.route()
-    }
-  }
-
-  /**
-   * Returns true if the host and port are unchanged from when this was created. This is used to
-   * detect if followups need to do a full connection-finding process including DNS resolution, and
-   * certificate pin checks.
-   */
-  fun sameHostAndPort(url: HttpUrl): Boolean {
-    val routeUrl = address.url
-    return url.port == routeUrl.port && url.host == routeUrl.host
   }
 }
